refactor(idp-basic): align package names with paths
