+++ plugins/idp-basic/src/test/java/org/apache/gravitino/idp/basic/storage/relational/mapper/provider/postgresql/TestIdpUserMetaPostgreSQLProvider.java
@@ -1,0 +1,42 @@
+/*
+ * Licensed to the Apache Software Foundation (ASF) under one
+ * or more contributor license agreements.  See the NOTICE file
+ * distributed with this work for additional information
+ * regarding copyright ownership.  The ASF licenses this file
+ * to you under the Apache License, Version 2.0 (the
+ * "License"); you may not use this file except in compliance
+ * with the License.  You may obtain a copy of the License at
+ *
+ *  http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing,
+ * software distributed under the License is distributed on an
+ * "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
+ * KIND, either express or implied.  See the License for the
+ * specific language governing permissions and limitations
+ * under the License.
+ */
+
+package org.apache.gravitino.idp.basic.storage.relational.mapper.provider.postgresql;
+
+import org.apache.gravitino.idp.basic.storage.relational.mapper.provider.base.IdpUserMetaBaseSQLProvider;
+import org.apache.gravitino.idp.basic.storage.relational.mapper.provider.base.TestIdpUserMetaBaseSQLProvider;
+
+public class TestIdpUserMetaPostgreSQLProvider extends TestIdpUserMetaBaseSQLProvider {
+
+  @Override
+  protected IdpUserMetaBaseSQLProvider createProvider() {
+    return new IdpUserMetaPostgreSQLProvider();
+  }
+
+  @Override
+  protected String expectedDeleteAtClause() {
+    return "deleted_at = CAST(EXTRACT(EPOCH FROM CURRENT_TIMESTAMP) * 1000 AS BIGINT)";
+  }
+
+  @Override
+  protected String expectedDeleteIdpUserMetasByLegacyTimelineSql() {
+    return "DELETE FROM idp_user_meta WHERE user_id IN (SELECT user_id FROM idp_user_meta"
+        + " WHERE deleted_at > 0 AND deleted_at < #{legacyTimeline} LIMIT #{limit})";
+  }
+}
