--- plugins/idp-basic/src/test/java/org/apache/gravitino/storage/relational/mapper/provider/h2/TestIdpUserMetaH2Provider.java
@@ -1,37 +1,0 @@
-/*
- * Licensed to the Apache Software Foundation (ASF) under one
- * or more contributor license agreements.  See the NOTICE file
- * distributed with this work for additional information
- * regarding copyright ownership.  The ASF licenses this file
- * to you under the Apache License, Version 2.0 (the
- * "License"); you may not use this file except in compliance
- * with the License.  You may obtain a copy of the License at
- *
- *  http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing,
- * software distributed under the License is distributed on an
- * "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
- * KIND, either express or implied.  See the License for the
- * specific language governing permissions and limitations
- * under the License.
- */
-
-package org.apache.gravitino.storage.relational.mapper.provider.h2;
-
-import org.apache.gravitino.storage.relational.mapper.provider.base.IdpUserMetaBaseSQLProvider;
-import org.apache.gravitino.storage.relational.mapper.provider.base.TestIdpUserMetaBaseSQLProvider;
-
-public class TestIdpUserMetaH2Provider extends TestIdpUserMetaBaseSQLProvider {
-
-  @Override
-  protected IdpUserMetaBaseSQLProvider createProvider() {
-    return new IdpUserMetaH2Provider();
-  }
-
-  @Override
-  protected String expectedDeleteAtClause() {
-    return "deleted_at = DATEDIFF('MILLISECOND', TIMESTAMP '1970-01-01 00:00:00',"
-        + " CURRENT_TIMESTAMP())";
-  }
-}
